requirements.txt
alembic==1.8.1
attrs==19.3.0
certifi==2022.6.15
chardet==3.0.4
click==8.1.3
dnspython==1.16.0
eventlet==0.30.2
Flask==2.0.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
gpxpy==1.4.2
greenlet==1.1.2
gunicorn==20.0.4
idna==2.8
importlib-metadata==4.12.0
importlib-resources==5.9.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
passlib==1.7.4
psycopg2==2.8.6
PyJWT==2.4.0
pyproj==3.3.1
python-dotenv==0.19.2
python-engineio==3.14.0
python-socketio==4.6.0
requests==2.22.0
Shapely==1.8.1.post1
shortuuid==1.0.9
simplekv==0.14.1
six==1.16.0
SQLAlchemy==1.4.32
urllib3==1.25.11
Werkzeug==2.0.3
XJ-Flask-JWT-Extension==0.0.2
zipp==3.8.1
geopy